requirements.txt
kafka-python==2.0.2
six>=1.16.0
pyspark==3.5.1
pandas==2.2.2
requests==2.32.2
Flask==3.0.3
Flask-Cors==4.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
flask-restful==0.3.10
matplotlib==3.8.4
streamlit
plotly
